requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
python-dotenv>=1.0.0

# Blockchain dependencies
web3>=6.0.0
eth-account>=0.9.0
py-solc-x>=1.12.0
# Note: ganache-cli is a Node.js package, install with: npm install -g ganache-cli